Adapt Setting to work with double type

+++ include/settings.hpp
@@ -54,7 +54,7 @@ public:
         std::cout << "Settings are loaded." << std::endl;
     }
 
-    int get_value(const std::string& key){
+    double get_value(const std::string& key){
         if (_settings.find(key) == _settings.end()){
             throw std::runtime_error("Key: " + key + " is not in settings");
         } else {
@@ -81,7 +81,7 @@ public:
     }
 
 private:
-    std::unordered_map<std::string, int> _settings;
+    std::unordered_map<std::string, double> _settings;
 
 };
